refactor: consolidate video processing utilities into videoDomUtils service
